@@ -1,48 +1,11 @@
-from typing import Any, Mapping
-
-import requests
-from requests.adapters import HTTPAdapter
-from urllib3.util import Retry
-
-DEFAULT_TIMEOUT = 10  # seconds
+import httpx
 
 
-class TimeoutHTTPAdapter(HTTPAdapter):
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        self.timeout = DEFAULT_TIMEOUT
-        if "timeout" in kwargs:
-            self.timeout = kwargs["timeout"]
-            del kwargs["timeout"]
-        super().__init__(*args, **kwargs)
+def get_client() -> httpx.AsyncClient:
+    client = httpx.AsyncClient()
 
-    def send(
-        self,
-        request: requests.PreparedRequest,
-        stream: bool = False,
-        timeout: None | float | tuple[float, float] | tuple[float, None] = None,
-        verify: bool | str = True,
-        cert: None | bytes | str | tuple[bytes | str, bytes | str] = None,
-        proxies: Mapping[str, str] | None = None,
-    ) -> requests.Response:
-        if timeout is None:
-            timeout = self.timeout
-        return super().send(request, stream, timeout, verify, cert, proxies)
+    client.headers.update(
+        {"user-agent": "ChronosEDT/1.0.0"}
+    )  # TODO: Use version in root __init__.py
 
-
-def get_client(
-    retries: int = 3, backoff_factor: int = 1, timeout: int = DEFAULT_TIMEOUT
-) -> requests.Session:
-    retry_strategy = Retry(
-        total=retries,
-        backoff_factor=backoff_factor,
-        status_forcelist=[429, 500, 502, 503, 504],
-    )
-    adapter = TimeoutHTTPAdapter(max_retries=retry_strategy, timeout=timeout)
-
-    http = requests.Session()
-    http.mount("https://", adapter)
-    http.mount("http://", adapter)
-
-    http.headers.update({"user-agent": "EDTVelizy/1.0.0"})
-
-    return http
+    return client
